feat: implement book copy management in BookCopyController; add getAllBookCopies endpoint and refactor inventory handling
