+++ inventory-service/src/main/java/com/example/inventoryservice/controller/BookCopyController.java
@@ -1,23 +1,80 @@
 package com.example.inventoryservice.controller;
 
+import com.example.inventoryservice.dto.BookCopyRequestDTO;
 import com.example.inventoryservice.entity.BookCopy;
+import com.example.inventoryservice.entity.Inventory;
+import com.example.inventoryservice.enums.Status;
 import com.example.inventoryservice.service.BookCopyService;
+import com.example.inventoryservice.service.InventoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.Map;
+import java.util.UUID;
 
 @RestController
 @RequestMapping("/api/v1/inventory-service/copies")
 public class BookCopyController {
-
+    @Autowired
+    private InventoryService inventoryService;
     @Autowired
     private BookCopyService bookCopyService;
 
     @PostMapping
-    public ResponseEntity<BookCopy> addBookCopy(@RequestBody BookCopy bookCopy) {
-        BookCopy savedBookCopy = bookCopyService.addBookCopy(bookCopy);
-        return ResponseEntity.ok(savedBookCopy);
+    public ResponseEntity<?> addBookCopy(@RequestBody BookCopyRequestDTO request) {
+        UUID bookId = request.getBookId();
+        int quantity = request.getQuantity();
+        String location = request.getLocation();
+
+        Inventory inventory = inventoryService.getInventoryByBookId(bookId);
+
+        if (inventory == null) {
+            Inventory newInventory = new Inventory();
+            newInventory.setAvailable(quantity);
+            newInventory.setTotalQuantity(quantity);
+            newInventory.setLost(0);
+            newInventory.setDamaged(0);
+            newInventory.setBorrowed(0);
+            newInventory.setBookId(bookId);
+            inventory = inventoryService.saveInventory(newInventory);
+        }
+
+        for (int i = 0; i < quantity; i++) {
+            BookCopy bookCopy = new BookCopy();
+            String latestCode = bookCopyService.findLatestCopyCode();
+            int nextNumber = 1;
+            if (latestCode != null && latestCode.startsWith("BC_")) {
+                nextNumber = Integer.parseInt(latestCode.substring(3)) + 1;
+            }
+            String formattedCode = String.format("BC_%03d", nextNumber);
+            bookCopy.setCopyCode(formattedCode);
+            bookCopy.setBookId(bookId);
+            bookCopy.setLocation(location);
+            bookCopy.setStatus(Status.AVAILABLE);
+            bookCopy.setInventory(inventory);
+            bookCopyService.addBookCopy(bookCopy);
+
+            inventory.setTotalQuantity(inventory.getTotalQuantity() + 1);
+            inventory.setAvailable(inventory.getAvailable() + 1);
+        }
+
+        inventoryService.saveInventory(inventory);
+
+        return ResponseEntity.ok(Map.of("message", "Added " + quantity + " book copies successfully"));
     }
+
+    @GetMapping
+    public ResponseEntity<?> getAllBookCopies() {
+        return ResponseEntity.ok(bookCopyService.getAllBookCopy());
+    }
+
+//    @PostMapping
+//    public ResponseEntity<BookCopy> addBookCopy(@RequestBody BookCopy bookCopy) {
+//        BookCopy savedBookCopy = bookCopyService.addBookCopy(bookCopy);
+//        return ResponseEntity.ok(savedBookCopy);
+//    }
 
     @GetMapping("/latest-copy-code")
     public ResponseEntity<String> getLatestCopyCode() {
