+++ inventory-service/src/main/java/com/example/inventoryservice/service/BookCopyService.java
@@ -2,7 +2,10 @@ package com.example.inventoryservice.service;
 
 import com.example.inventoryservice.entity.BookCopy;
 
+import java.util.List;
+
 public interface BookCopyService {
     BookCopy addBookCopy(BookCopy bookCopy);
+    List<BookCopy> getAllBookCopy();
     String findLatestCopyCode();
 }
